requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
ao3-api==2.3.0
    # via -r requirements.in
asgiref==3.5.2
    # via django
asttokens==2.0.8
    # via stack-data
attrs==22.1.0
    # via pytest
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via ao3-api
black==22.10.0
    # via -r requirements.in
build==0.8.0
    # via pip-tools
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
coverage[toml]==6.5.0
    # via pytest-cov
decorator==5.1.1
    # via ipython
dirty-equals==0.5.0
    # via -r requirements.in
dj-database-url==1.0.0
    # via environs
dj-email-url==1.0.6
    # via environs
dj-rest-auth==2.2.5
    # via -r requirements.in
django==4.1
    # via
    #   -r requirements.in
    #   dj-database-url
    #   dj-rest-auth
    #   django-cors-headers
    #   django-db-geventpool
    #   django-extensions
    #   django-filter
    #   django-taggit
    #   djangorestframework
    #   model-bakery
django-cache-url==3.4.2
    # via environs
django-cors-headers==3.13.0
    # via -r requirements.in
django-db-geventpool==4.0.1
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-taggit==3.0
    # via -r requirements.in
django-test-plus==2.2.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   dj-rest-auth
environs[django]==9.5.0
    # via -r requirements.in
execnet==1.9.0
    # via pytest-xdist
executing==1.1.1
    # via stack-data
gevent==22.10.1
    # via -r requirements.in
greenlet==1.1.3.post0
    # via gevent
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
iniconfig==1.1.1
    # via pytest
ipython==8.5.0
    # via -r requirements.in
isort==5.10.1
    # via -r requirements.in
jedi==0.18.1
    # via ipython
lxml==4.9.1
    # via ao3-api
marshmallow==3.18.0
    # via environs
matplotlib-inline==0.1.6
    # via ipython
model-bakery==1.8.0
    # via -r requirements.in
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via
    #   build
    #   marshmallow
    #   pytest
parso==0.8.3
    # via jedi
pathspec==0.10.1
    # via black
pep517==0.13.0
    # via build
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.9.0
    # via -r requirements.in
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.31
    # via ipython
psycogreen==1.0.2
    # via django-db-geventpool
psycopg2==2.9.4
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via
    #   pytest
    #   pytest-forked
pygments==2.13.0
    # via ipython
pyparsing==3.0.9
    # via packaging
pytest==7.1.3
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-forked
    #   pytest-xdist
pytest-cov==4.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-forked==1.4.0
    # via pytest-xdist
pytest-xdist==2.5.0
    # via -r requirements.in
python-dotenv==0.21.0
    # via environs
pytz==2022.5
    # via
    #   dirty-equals
    #   djangorestframework
requests==2.28.1
    # via
    #   ao3-api
    #   responses
responses==0.22.0
    # via -r requirements.in
six==1.16.0
    # via asttokens
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
stack-data==0.5.1
    # via ipython
toml==0.10.2
    # via responses
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   pep517
    #   pytest
traitlets==5.5.0
    # via
    #   ipython
    #   matplotlib-inline
types-toml==0.10.8
    # via responses
typing-extensions==4.4.0
    # via dirty-equals
urllib3==1.26.12
    # via
    #   requests
    #   responses
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.1
    # via pip-tools
whitenoise==6.2.0
    # via -r requirements.in
zope-event==4.5.0
    # via gevent
zope-interface==5.5.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
